Finalization of OC module
This commit also changes around the PPS file. PPS is now defined simply as a pin, because there are no in or output pin specifics
The OC has also been modified. It no longer accepts a callback funtion. This is due to the fact that an OC interrupt is not fired
when the PWM wave form is generated. Users should use the associated timer callback function instead.