Partial Correct of Serial C
This correction is untested. The addition involves addition of two
file handles. One handle is used for reading and the other for writing.

Serial C++ is untested as of yet, and will most likely not work.
The reason it will probably not work is because the read() call returns as soon
as a single piece of data is read off the serial port. It will then return
the number of data bytes that it was able to read. The read command must wait
to be executed until a sufficient amount of data is read from the port.