Addition of program_linker.ld script
This script is now standalone so it can be included in all projects.
This maps the memory to appropriate locations within the microcontroller
so that the program may be loaded with the bootloader. NOTE: Programs will
not function normally when this linker script is active! It must work
in conjunction with the bootloader.